✅ Fix testGetPseudoRandomNumberFuzz test

+++ test/foundry/utils/TestHelpersTests.t.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.13;
 import {TestHelpers} from "../utils/TestHelpers.t.sol";
 
 contract TestHelpersTests is TestHelpers {
-    function testGetPseudoRandomNumberFuzz(uint256 max, uint256 min, uint256 salt) public {
+    function testGetPseudoRandomNumberFuzz(uint256 max, uint16 min, uint128 salt) public {
         vm.assume(max >= min);
         uint256 result = getPseudoRandomNumber(max, min, salt);
         assertTrue(result >= min && result <= max);
